+++ app/src/renderer/components/common/TmPage.vue
@@ -8,7 +8,7 @@
         {{ subtitle }}
       </h3>
       <slot slot="menu-body" name="menu-body">
-        <tm-balance v-if="user.signedIn" />
+        <tm-balance v-if="session.signedIn" />
         <tool-bar :refresh="refreshable" :searching="searchable" />
       </slot>
       <slot slot="header-buttons" name="header-buttons" />
@@ -113,7 +113,7 @@ export default {
   },
   data: () => ({ ps: `` }),
   computed: {
-    ...mapGetters([`user`, `filters`, `connected`]),
+    ...mapGetters([`session`, `filters`, `connected`]),
     searchable({ somethingToSearch, setSearch } = this) {
       return this.search ? { somethingToSearch, setSearch } : undefined
     },